added admin/ to url, send_mail replaced to service
