+++ apps/core/utils.py
@@ -2,12 +2,9 @@ import logging
 import urllib
 from functools import wraps
 
-from anymail.exceptions import AnymailConfigurationError, AnymailInvalidAddress, AnymailRequestsAPIError
 from django.conf import settings
 from django.core.files.base import ContentFile
-from django.core.mail import EmailMessage
 from django.template.defaultfilters import slugify as django_slugify
-from rest_framework import status
 from rest_framework.response import Response
 
 from apps.core.constants import ALPHABET, STATUS_INFO
@@ -133,39 +130,6 @@ def get_app_list(self, request):
             app["models"].sort(key=lambda x: x["name"])
 
     return app_list
-
-
-def send_email(from_email: str, to_emails: tuple, template_id: str, context: dict, attach_file: bool = False) -> bool:
-    """Transmit email context in the template and send email.
-
-    If there is a file in the instance then attach_file=True.
-
-    For more information please read the documentation:
-    https://anymail.readthedocs.io/en/stable/esps/mailjet//
-    """
-    try:
-        message = EmailMessage(from_email=from_email, to=to_emails)
-        message.template_id = template_id
-        message.merge_global_data = context
-        if attach_file:
-            file_path = context.get("file_path")
-            message.attach_file(file_path)
-        message.send()
-        if hasattr(message, "anymail_status") and message.anymail_status.esp_response.status_code == status.HTTP_200_OK:
-            return True
-        return False
-    except AnymailConfigurationError as error:
-        msg = f"Неверные настройки Mailjet. Не удалось отправить письмо на почтовые адреса - {to_emails}."
-        logger.critical(msg, error, exc_info=True)
-    except (AnymailRequestsAPIError, AnymailInvalidAddress) as error:
-        msg = (
-            f"Не указан адрес электронной почты отправителя."
-            f"Не удалось отправить письмо на почтовые адреса - {to_emails}."
-        )
-        logger.critical(msg, error, exc_info=True)
-    except ValueError as error:
-        msg = f"Неверный ID шаблона Mailjet. Не удалось отправить письмо на почтовые адреса - {to_emails}."
-        logger.critical(msg, error, exc_info=True)
 
 
 def get_domain(request):
