+++ apps/info/views/question.py
@@ -5,7 +5,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from apps.core.models import Setting
-from apps.core.utils import send_email
+from apps.core.services.send_email import send_email
 from apps.info.models import Question
 from apps.info.schema.schema_extension import ERROR_MESSAGES_FOR_QUESTION_FOR_400
 
@@ -14,11 +14,9 @@ class QuestionCreateAPIView(APIView):
     """Create Question."""
 
     class QuestionSerializer(serializers.ModelSerializer):
-        sent = serializers.CharField(read_only=True)
-
         class Meta:
             model = Question
-            fields = "__all__"
+            exclude = ("sent",)
 
     @extend_schema(
         request=QuestionSerializer,
@@ -29,21 +27,21 @@ class QuestionCreateAPIView(APIView):
     )
     def post(self, request):
         serializer = self.QuestionSerializer(data=request.data)
-        if serializer.is_valid():
-            instance = serializer.save()
-            from_email = Setting.get_setting("email_send_from")
-            to_emails = (Setting.get_setting("email_to_send_questions"),)
-            template_id = settings.MAILJET_TEMPLATE_ID_QUESTION
-            context = {
-                "question": instance.question,
-                "author_name": instance.author_name,
-                "author_email": instance.author_email,
-            }
+        if not serializer.is_valid():
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        instance = serializer.save()
+        from_email = Setting.get_setting("email_send_from")
+        to_emails = (Setting.get_setting("email_to_send_questions"),)
+        template_id = settings.MAILJET_TEMPLATE_ID_QUESTION
+        context = {
+            "question": instance.question,
+            "author_name": instance.author_name,
+            "author_email": instance.author_email,
+        }
 
-            response_success = send_email(from_email, to_emails, template_id, context)
+        response_success = send_email(from_email, to_emails, template_id, context)
 
-            if response_success:
-                instance.sent = True
-                instance.save()
-            return Response(status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if response_success:
+            instance.sent = True
+            instance.save()
+        return Response(status=status.HTTP_201_CREATED)
