+++ apps/users/models.py
@@ -6,8 +6,6 @@ from django.utils import timezone
 
 class UserManager(BaseUserManager):
     def _create_user(self, email, password, is_staff, is_superuser, **extra_fields):
-        if not email:
-            raise ValueError("Users must have an email address")
         now = timezone.now()
         email = self.normalize_email(email)
         user = self.model(
@@ -28,6 +26,8 @@ class UserManager(BaseUserManager):
 
     def create_superuser(self, email, password, **extra_fields):
         user = self._create_user(email, password, True, True, **extra_fields)
+        user.first_name = user.username
+        user.save()
         return user
