--- templates/admin/auth/user/add_form.html
@@ -1,12 +1,0 @@
-{% extends "admin/change_form.html" %}
-{% load i18n %}
-
-{% block form_top %}
-  {% if not is_popup %}
-    <p>Введите имя, фамилию пользователя и адрес его электронной почты, нажмите "Сохранить".
-      На указанный почтовый ящик пользователю придет ссылка,
-      по которой он сможет зайти на сайт и  установить свой пароль.</p>
-  {% else %}
-    <p>Введите имя пользователя и его электронный почтовый ящик.</p>
-  {% endif %}
-{% endblock %}
